Fix scoring unit tests

+++ test/ai/move-effect-scores/hazards.test.ts
@@ -23,7 +23,7 @@ describe("Move Effect Scores - Hazards", () => {
   beforeEach(() => {
     game = new GameManager(phaserGame);
     game.override
-      .battleType("double")
+      .battleType("single")
       .enemySpecies(SpeciesId.MAGIKARP)
       .enemyAbility(AbilityId.BALL_FETCH)
       .ability(AbilityId.BALL_FETCH)
